requirements.txt
APScheduler==3.6.3
CacheControl==0.13.1
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
docopt==0.6.2
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
pipreqs==0.4.13
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python-telegram-bot==13.13
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
tornado==6.3.2
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
yarg==0.1.9
